Mentor Create Page
Added a BASIC template for the Mentor Create Page in my own files that I created:
1. MentorCreatePage.css
2. MentorCreatePage.jsx

I was thinking that if we like these as a group, we can copy code over into the EXISTING FILES app.css and app.jsx files for a 'General Template' and can then later edit the css to style every new page differently. For now, these files are going up to GIT and I can merge them into app.css and app.jsx after frontend discussion meeting if we think they are applicable enough.