Edited the ProgressBar and implemented them into ProgramDetailPage & SessionDetailPage
Please note: No edits made to PageContent that affect functionality. Upon saving the page my IDE made some spacing changes and changed ' to "" and I just left it.

Edited the ProgressBar code to make it custom for this project and implemented it into the ProgramDetailPage & the SessionDetailPage. Have also added 'PageContent' to SessionDetailPage to make it more consistent.