Fix for the ProgressBar component - Handling rounding up
There was an issue where the percentage on the ProgressBar wasn't getting handled properly and was spitting out detailed calculations. I have implemented the 'Math.ceil' method to handle this.

As requested by the team, I have also un-commented out some buttons on the App.jsx file to make them functional as an addition.