fix: keep footer at bottom of page
Remove the `height` property from the `#root` element and add the `min-height` property to the `body` element in the `index.css` file to allow the body to expand to fit the content. this allows `Footer.css` file to position the footer at the bottom of the content area using `position: absolute` and `bottom: 0` and `margin-top: auto`. Set `min-height: 100vh` for the `body` element to ensure that the body is at least as tall as the viewport.