ProgramDetail and MentorDetail Styling
Amendments have been made to the ProgramDetail and MentorDetail pages for styling.

Something was wrong with the progressbar component on the MentorDetail page and so extra code was handled to fix this. I had to force a width on the progressbar and wrap it in a <div> that uses negative margin to edit the container white. It's abit of a hack but it works. This is because the table elements in this code restricted the logic for my progressbar component and this was the solution to fix it. The table elements were not allowing the progressbar to render properly without it.

I have also added a new button component to the SessionDetailPage as required to link a page where we can edit the SessionDetailpage.